Improve the faraday cache middleware
